- add Streamlit UI

+++ src/sahi_tracking/experiments_framework/DataStatePersistance.py
@@ -8,15 +8,18 @@ from sahi_tracking.helper.config import get_local_data_path
 
 
 class DataStatePersistance:
-    def __init__(self):
+    def __init__(self, read_only= False):
         self.local_data_path =  get_local_data_path()
         self.state_path: Path = self.local_data_path / 'state.pkl'
+        self.read_only = read_only
 
         self.state = None
         self.load_state()
+
     def write_state(self):
-        with open(self.state_path, 'wb') as handle:
-            pickle.dump(self.state, handle)
+        if not self.read_only:
+            with open(self.state_path, 'wb') as handle:
+                pickle.dump(self.state, handle)
 
 
     def load_state(self):
@@ -25,6 +28,8 @@ class DataStatePersistance:
                 self.state = pickle.load(handle)
         else:
             print("state_path does not exist. Creating new state.")
+            import time
+            time.sleep(30)
             self.create_new_state()
             self.write_state()
 
@@ -38,14 +43,19 @@ class DataStatePersistance:
         }
         self.write_state()
 
-    def delete_existing(self, key, hash):
+    def delete_all_by_key(self, key):
+        self.load_state()
+        self.state[key] = []
+        self.write_state()
+
+
+    def delete_existing_by_hash(self, key, hash):
         self.load_state()
         print("Deleting existing")
 
         self.state[key] = [item for item in self.state[key] if item["hash"] != hash]
         assert not hash in [item for item in self.state[key]]
         self.write_state()
-
 
     def update_state(self, type, key, value):
         self.load_state()
@@ -64,3 +74,13 @@ class DataStatePersistance:
         self.load_state()
         return hash in [item["hash"] for item in self.state[key]]
 
+
+if __name__ == "__main__":
+    parser = ArgumentParser()
+    parser.add_argument("action", choices=['delete_all_by_key'])
+    parser.add_argument("--key" , choices=['datasets', 'tracking_results', 'predictions_results', 'evaluation_results'])
+    args = parser.parse_args()
+
+    if args.action == "delete_all_by_key":
+        state = DataStatePersistance()
+        state.delete_all_by_key(args.key)
